+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 
+### Changed
+- 복구 Import/Preview의 임시 SwiftData store 수명과 cleanup 경계를 정리해 QA/test 로그의 sqlite 임시 파일 경고를 제거.
+
 ## [2.3.10] - 2026-05-05
 
 ### Changed